refactor: split RPCHandlers.swift into focused handler groups (CROW-1134)
makeCommandRouter is now a thin assembler that merges per-concern maps.
The Linux parity gate globs every *RPCHandlers.swift file so groups can
leave the original 3k-line router without going blind to ADR 0016.

Crow-Session: 7CB7D345-523E-4D47-8371-58CEA580EAA3
